==> notebooks/00_setup.ipynb <==
#### Imports & Constants

import pathlib
import contextlib
import pandas as pd

pd.set_option('display.max_columns', 110) # 110 is the number of columns in the dataset
pd.set_option('display.max_rows', 1000)

DATASET_PATH = pathlib.Path("..") / "data" / "raw" / "players.csv"
DATASET_INFO_PATH = pathlib.Path("..") / "data" / "raw" / "info.txt"

#### Create info.txt

df = pd.read_csv(DATASET_PATH, low_memory=False)

with open("../data/raw/info.txt", "w") as f:
    with contextlib.redirect_stdout(f):
        df.info(verbose=True, show_counts=True)

#### Sample Data

df.head()

df.describe()
